create cli module
